khkh

+++ server/ride_sharing_application/RideSharing.API/Models/Domain/City.cs
@@ -24,8 +24,8 @@ namespace RideSharing.API.Models.Domain
         [MaxLength(10)]
         public string? Pincode { get; set; }
         
-        public double? Latitude { get; set; }
-        public double? Longitude { get; set; }
+        public float? Latitude { get; set; }
+        public float? Longitude { get; set; }
         
         public bool IsActive { get; set; } = true;
